+++ src/main/java/lagatrix/client/gui/views/main/form/MainView.java
@@ -2,6 +2,7 @@ package lagatrix.client.gui.views.main.form;
 
 import lagatrix.client.gui.views.main.getters.Getter;
 import javax.swing.JPanel;
+import lagatrix.client.connection.RequesterManager;
 
 /**
  * This class represents the views of the main program, initialises the getter 
@@ -12,7 +13,8 @@ import javax.swing.JPanel;
  */
 public abstract class MainView extends JPanel {
     
-    public Getter getter;
+    protected Getter getter;
+    protected RequesterManager requester;
 
     /**
      * Constructor of the class.
@@ -26,6 +28,15 @@ public abstract class MainView extends JPanel {
      * @return The class instancied.
      */
     public abstract Getter inicialiceGetter();
+    
+    /**
+     * Set the requester to make calls.
+     * 
+     * @param requester The requester.
+     */
+    public void setRequester(RequesterManager requester) {
+        this.requester = requester;
+    }
     
     /**
      * Start the view.