+++ src/main/java/lagatrix/client/gui/views/main/getters/Getter.java
@@ -1,35 +1,48 @@
 package lagatrix.client.gui.views.main.getters;
 
+import lagatrix.client.connection.RequesterManager;
+import lagatrix.client.exceptions.BadExecutionException;
+import lagatrix.client.exceptions.connection.ConnectionException;
 import lagatrix.client.gui.views.main.form.MainView;
 
 /**
- * This class obtain the infornmation who see in the views of the main views.
+ * This class obtain the infornmation who see in the views of the main views 
+ * over a period of time.
  *
  * @author javierfh03
  * @since 0.2
  */
 public abstract class Getter extends Thread{
     
-    public boolean run;
-    public int pause;
-    public MainView view;
+    private boolean run;
+    private int pause;
+    protected RequesterManager requester;
+    protected MainView view;
     
     /**
      * The constructor of the class.
      * 
      * @param view The view who assign the getter.
+     * @param requester The requester.
      * @param pause The time between one application and another.
      */
-    public Getter(MainView view, int pause){
+    public Getter(MainView view, RequesterManager requester, int pause){
         this.view = view;
         this.run = true;
+        this.requester = requester;
         this.pause = pause;
     }
 
     @Override
-    public final void run(){
+    public final void run() {
         while (run) {            
-            getsInformation();
+            try {
+                getsInformation();
+            } catch (ConnectionException ex) {
+                
+            } catch (BadExecutionException ex) {
+                
+            }
             try {
                 sleep(pause);
             } catch (InterruptedException ex) {
@@ -40,8 +53,11 @@ public abstract class Getter extends Thread{
     
     /**
      * Put in the components the information of the server.
+     * 
+     * @throws ConnectionException If have a error in the connection.
+     * @throws BadExecutionException If have a error in the execution of action.
      */
-    public abstract void getsInformation();
+    public abstract void getsInformation() throws ConnectionException, BadExecutionException;
     
     /**
      * Kill the thread.