+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LACLI003 - 24/04/2023 - Add the Getter and MainView class.
 
 ### Changed
+- LACLI003 - 25/04/2023 - Add forgotten things in Getter and MainView class.
 
 ### Fixed